AI 功能上线

+++ internal/ai/internal/domain/jd.go
@@ -7,8 +7,8 @@ const (
 )
 
 type JDEvaluation struct {
-	Score    int    `json:"score"`
-	Analysis string `json:"analysis"`
+	Score    float64 `json:"score"`
+	Analysis string  `json:"analysis"`
 }
 
 type JD struct {
